requirements.txt
flask==3.0.0
reportlab>=4.0.0
werkzeug>=3.0.0
python-dateutil>=2.8.2
-e .